introduced editor structure
