+++ app/_costants.py
@@ -3,3 +3,4 @@ SCR_W = 1024
 SCR_H = 640
 RECT_W = 64
 RECT_H = 64
+CAM_SCALE = 1.6
