+++ app/game/__init__.py
@@ -1,7 +1,7 @@
+from .game import Game
 from .camera import Camera
-from .level import Level
 
 __all__ = [
-    'Camera',
-    'Level'
+    'Game',
+    'Camera'
 ]
