+++ app/level_editor/__init__.py
@@ -1,0 +1,5 @@
+from .editor import Editor
+
+__all__ = [
+    'Editor'
+]
