+++ main.py
@@ -2,4 +2,4 @@ import app
 
 
 if __name__ == "__main__":
-    app.Main().run()
+    app.Main()
